genetlink: use errors.Is where possible
