+++ go.mod
@@ -1,10 +1,10 @@
 module github.com/mdlayher/genetlink
 
-go 1.12
+go 1.13
 
 require (
 	github.com/google/go-cmp v0.3.1
-	github.com/mdlayher/netlink v0.0.0-20190828143259-340058475d09
-	golang.org/x/net v0.0.0-20190827160401-ba9fcec4b297
-	golang.org/x/sys v0.0.0-20190826190057-c7b8b68b1456
+	github.com/mdlayher/netlink v0.0.0-20191004170026-3c8695cb0643
+	golang.org/x/net v0.0.0-20191003171128-d98b1b443823
+	golang.org/x/sys v0.0.0-20191003212358-c178f38b412c
 )
